+++ docs/en-US/guide/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.0.61
+
+- 添加 LivePhoto 组件及其文档，支持视频和图片的长按播放功能 ([ddc62ed](https://github.com/iceywu/uni-ui-plus/commit/ddc62ed))
+- Add LivePhoto component and enhance documentation ([cbfacd0](https://github.com/iceywu/uni-ui-plus/commit/cbfacd0))
+- 更新 LivePhoto 组件的指示器样式，添加纯 CSS 图标支持 ([7379731](https://github.com/iceywu/uni-ui-plus/commit/7379731))
+
 ## v0.0.60
 
 [compare changes](https://github.com/iceywu/uni-ui-plus/compare/v0.0.58...v0.0.60)